+++ lib/data/local/comic.dart
@@ -6,6 +6,9 @@ class ComicItem {
   String? thumbnail;
   String? uriDetails;
 
+  static const defaultThumbnailUrl =
+      'https://static.posters.cz/image/750/plakaty/marvel-comic-here-come-the-heroes-i34927.jpg';
+
   ComicItem(
       {this.id,
       this.title,
