--- lib/features/dashboard/cubit/comic_state.dart
@@ -1,6 +1,0 @@
-part of 'comic_cubit.dart';
-
-@immutable
-abstract class ComicState{}
-
-class ComicInitial extends ComicState {}
