+++ lib/features/dashboard/dashboard_module.dart
@@ -1,12 +1,10 @@
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:marvel_app/features/dashboard/bloc/comic_search_bloc.dart';
-import 'package:marvel_app/features/dashboard/cubit/comic_cubit.dart';
-
 import 'package:marvel_app/utils/navigation_paths.dart';
-
 import '../dashboard/home/home_page.dart';
 import '../dashboard/search/search_page.dart';
 import 'detail/detail_page.dart';
+import 'home/cubit/comic_cubit.dart';
+import 'search/bloc/comic_search_bloc.dart';
 
 class DashboardModule extends Module {
   @override
