+++ lib/features/dashboard/search/bloc/comic_search_event.dart
