+++ lib/features/dashboard/search/bloc/comic_search_state.dart
